+++ packages/block-library/src/utils/get-template-part-area-icon.js
@@ -7,13 +7,13 @@ import {
 } from '@wordpress/icons';
 
 /**
- * Helper function to retrieve the corresponding icon by area name.
+ * Returns the icon of a template part area.
  *
  * @param {string} areaOrIconName The area name (e.g., 'header', 'navigation-overlay').
  *
  * @return {Object} The corresponding icon.
  */
-export const getTemplatePartIcon = ( areaOrIconName ) => {
+export const getTemplatePartAreaIcon = ( areaOrIconName ) => {
 	if ( 'header' === areaOrIconName ) {
 		return headerIcon;
 	} else if ( 'footer' === areaOrIconName ) {